Remove explicit setting of dispatcher
Older versions of newrelic_rpm required setting the dispatcher for the agent to
report. This hasn't been the case for most of the last year, though.

Setting the dispatcher on manual_start wouldn't be a problem except DelayedJob
(and potentially Resque) are 1) run through rake and 2) rely on the background
framework being detected as dispatcher to function properly.

Explicitly setting the `:dispatcher => :rake` short-cuts that detection, and
things break. Removing that dispatcher will only make processes on the
environment page not list 'rake' preceding the hostname, which is a bummer, but
better than DJ clients missing instrumentation.